Move basket API to core app
