+++ doc/api/shuup.core.rst
@@ -7,6 +7,7 @@ Subpackages
 .. toctree::
 
     shuup.core.api
+    shuup.core.basket
     shuup.core.cache
     shuup.core.defaults
     shuup.core.fields
