--- doc/api/shuup.front.models.rst
@@ -1,22 +1,0 @@
-shuup.front.models package
-==========================
-
-Submodules
-----------
-
-shuup.front.models.stored_basket module
----------------------------------------
-
-.. automodule:: shuup.front.models.stored_basket
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: shuup.front.models
-    :members:
-    :undoc-members:
-    :show-inheritance:
