+++ shuup/front/admin_module/carts/views/_list.py
@@ -18,7 +18,7 @@ from shuup.admin.utils.picotable import (
 )
 from shuup.admin.utils.views import PicotableListView
 from shuup.core.models import Shop
-from shuup.front.models import StoredBasket
+from shuup.core.models import StoredBasket
 from shuup.utils.i18n import format_money, get_locally_formatted_datetime
