+++ shuup/front/apps/saved_carts/views.py
@@ -15,7 +15,7 @@ from django.views.generic.detail import SingleObjectMixin
 from shuup.core.excs import ProductNotOrderableProblem
 from shuup.core.models import OrderLineType, Product
 from shuup.core.utils.users import real_user_or_none
-from shuup.front.models import StoredBasket
+from shuup.core.models import StoredBasket
 from shuup.front.views.dashboard import DashboardViewMixin
