+++ shuup/front/middleware.py
@@ -20,7 +20,7 @@ from shuup.core.middleware import ExceptionMiddleware
 from shuup.core.models import (
     Contact, get_company_contact, get_person_contact, Shop
 )
-from shuup.front.basket import get_basket
+from shuup.core.basket import get_basket
 
 __all__ = ["ProblemMiddleware", "ShuupFrontMiddleware"]
 
@@ -60,7 +60,7 @@ class ShuupFrontMiddleware(object):
           ``request.person`` is a
           :class:`~shuup.core.models.PersonContact`.
 
-      ``request.basket`` : :class:`shuup.front.basket.objects.BaseBasket`
+      ``request.basket`` : :class:`shuup.core.basket.objects.BaseBasket`
           Shopping basket in use.
     """
