--- shuup/front/models/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-# This file is part of Shuup.
-#
-# Copyright (c) 2012-2017, Shoop Commerce Ltd. All rights reserved.
-#
-# This source code is licensed under the OSL-3.0 license found in the
-# LICENSE file in the root directory of this source tree.
-
-from .stored_basket import StoredBasket
-
-__all__ = ["StoredBasket"]
