+++ shuup/front/signals.py
@@ -7,8 +7,6 @@
 # LICENSE file in the root directory of this source tree.
 from django.core.signals import Signal
 
-# Modifying signals
-get_basket_command_handler = Signal(providing_args=["command"], use_caching=True)
 
 # Completion signals
 order_complete_viewed = Signal(providing_args=["order", "request"], use_caching=True)
