+++ shuup/testing/factories.py
@@ -41,8 +41,8 @@ from shuup.core.models import (
 from shuup.core.order_creator import OrderCreator, OrderSource
 from shuup.core.pricing import get_pricing_module
 from shuup.core.shortcuts import update_order_line_from_product
+from shuup.core.models import StoredBasket
 from shuup.default_tax.models import TaxRule
-from shuup.front.models import StoredBasket
 from shuup.testing.text_data import random_title
 from shuup.utils.filer import filer_image_from_data
 from shuup.utils.money import Money
