+++ shuup_tests/campaigns/test_basket_campaign_admin.py
@@ -5,7 +5,6 @@
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
 # test that admin actually saves catalog
-import decimal
 import pytest
 
 from django.test import override_settings
