+++ shuup_tests/campaigns/test_caching.py
@@ -25,7 +25,6 @@ def test_filter_caching(rf):
     catalog_filter = ProductFilter.objects.create()
     catalog_filter.products.add(product)
 
-
     catalog_campaign = CatalogCampaign.objects.create(shop=shop, active=True, name="test")
     catalog_campaign.filters.add(catalog_filter)
     assert CatalogFilterCachedShopProduct.objects.count() == 1
