+++ shuup_tests/campaigns/test_filters.py
@@ -14,7 +14,7 @@ from shuup.campaigns.models.catalog_filters import (
 )
 from shuup.campaigns.models.product_effects import ProductDiscountPercentage
 from shuup.core.models import Category, ProductMode, ShopProduct
-from shuup.front.basket import get_basket
+from shuup.core.basket import get_basket
 from shuup.testing.factories import (
     create_product, get_default_category, get_default_supplier,
     get_shipping_method
